почистил код: перенёс функции из файла carInfo/index в carInfo/utilities
