+++ src/components/admin/cars/index.jsx
@@ -20,13 +20,11 @@ class Cars extends Component {
 
   componentDidMount() {
     axios.get(`https://api.rent-auto.biz.tm/info_models`, getHeaders()).then(res => {
-    console.log(res.data)
     const list = res.data.sort((a,b) => (a.id > b.id) ? 1 : ((b.id > a.id) ? -1 : 0)); 
       this.setState({ cars: list, isLoading: false });
     });
   }
   render() {
-    console.log(this.state.cars);
     return (
       <AdminLayout>
         <div>
